# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.CreateEntityRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
    required:
    - fields
    type: object
...
  models.EntityListResponse:
    properties:
      entities:
        items:
          $ref: '#/definitions/models.EntityResponse'
        type: array
    type: object
  models.EntityResponse:
    properties:
      created_at:
        example: "2023-12-31T23:59:59Z"
        type: string
      deleted_at:
        example: "2023-12-31T23:59:59Z"
        type: string
      entity_id:
        example: entity_123
        type: string
      fields:
        additionalProperties: true
        type: object
      is_deleted:
        example: false
        type: boolean
      table_id:
        example: users_table
        type: string
      timestamp:
        example: "2023-12-31T23:59:59Z"
        type: string
    type: object
  models.ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
    type: object
...
  models.UpdateEntityRequest:
    properties:
      fields:
        additionalProperties: true
        type: object
    required:
    - fields
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the Notably API server for managing time-series data and entities.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Notably API
  version: "1.0"
paths:
...
  /tables/{tableId}/entities:
    get:
      consumes:
      - application/json
      description: Get a list of all entities in the specified table with optional
        time-based filtering
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Point-in-time query (RFC3339 format)
        in: query
        name: asOf
        type: string
      - description: Maximum number of entities to return
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of entities
          schema:
            $ref: '#/definitions/models.EntityListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: List entities in a table
      tags:
      - entities
    post:
      consumes:
      - application/json
      description: Create a new entity in the specified table
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Entity creation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateEntityRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Entity created successfully
          schema:
            $ref: '#/definitions/models.EntityResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new entity
      tags:
      - entities
  /tables/{tableId}/entities/{entityId}:
    delete:
      consumes:
      - application/json
      description: Mark an entity as deleted (soft delete)
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Entity deleted successfully
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Delete entity
      tags:
      - entities
    get:
      consumes:
      - application/json
      description: Retrieve a specific entity by its ID with optional point-in-time
        query
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      - description: Point-in-time query (RFC3339 format)
        in: query
        name: asOf
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Entity information
          schema:
            $ref: '#/definitions/models.EntityResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get entity by ID
      tags:
      - entities
    put:
      consumes:
      - application/json
      description: Update an existing entity's fields
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      - description: Entity update details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateEntityRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Entity updated successfully
          schema:
            $ref: '#/definitions/models.EntityResponse'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Update entity
      tags:
      - entities
  /tables/{tableId}/entities/{entityId}/history:
    get:
      consumes:
      - application/json
      description: Retrieve the history of changes for a specific entity
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      - description: Maximum number of history entries to return
        in: query
        name: limit
        type: integer
      - description: Start date for history (RFC3339 format)
        in: query
        name: since
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Entity history
          schema:
            $ref: '#/definitions/models.EntityListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Get entity history
      tags:
      - entities
  /tables/{tableId}/entities/{entityId}/undelete:
    post:
      consumes:
      - application/json
      description: Restore a previously deleted entity
      parameters:
      - description: Table ID
        in: path
        name: tableId
        required: true
        type: string
      - description: Entity ID
        in: path
        name: entityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Entity undeleted successfully
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Entity not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - Bearer: []
      summary: Undelete entity
      tags:
      - entities
...
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"